Add files via upload
I noticed that there is a discrepancy with Dr. L's document. It says that "106 is supposed to be secured in this section...", but this note used to be in SCP-513's chamber in the HCZ, and in 1.6 the chamber and the note were moved to the LCZ. So I decided to cover the inscription with blood.