Update loading_screens.jsonc
The full block (U+2588) in the SCP description on loading screens looks out of place, as it stands out too much, Imo. I decided to change it to "DATA REDACTED". Also changed "holding" in the description of the duck with the saxophone to "with", as in the game the duck doesn't hold a saxophone.