Editing the Concentration Camp Note, SCP-914 Test Log, and My Suggestions
So, new SCP-085 notes have been added to the game, which means now I need to change the code for Dr. Gears' office, and also make the very fine radio give the code for some other door instead of Dr. Gears' office. I added a number to the note from the concentration camp (859305, this is the prisoner number that can be seen on his texture). Since there are 6 digits there, and the code for the door is only 4, then two random digits out of 6 should either be splattered with blood or crossed out using the game code (for example, a random code appears on the SCP-372 document, so I think making the random two digits covered by overlay text or blood is not a problem). I calculated that there will be 15 possible combinations of the code, so it is unlikely that players will be able to remember them all and will have to touch SCP-1123 to find out the code.

Also, I slightly modified one of SCP-914's test logs, adding an experiment to upgrade the radio, so players know that the radio needs to be upgraded to very fine to get the access code (this will make the radio a useful item). What door code can the radio give? Well, I suggest making the elevator to some SCP that requires a code to be activated, making it initially non-functional until the player finds and enters the correct code (SCP-008, for example, which can be justified as an additional security measure, or SCP-457). These are just my suggestions, and do not necessarily need to be added to the game. Feel free to not accept the pull request. But I tried to explain why it should be done, since my original idea with SCP-085's notes, which are added to the game, was complex and implied the above changes.