Replacement of SCP-990 loading screen message about radio
This message is pretty useless, considering that the radio only gives the code to the door to Gears' office, which is not, in my opinion, enough to deserve a whole hint. In the previous request, I changed one of SCP-914's test logs, where I added the test with the radio, which is, in my opinion, a more logical and accessible hint. The new hint "The Red Ice does not care if you are a man or a bloodthirsty beast" concerns a more important gameplay mechanic, which is also not obvious to many players - the ability to kill hostile NPCs with SCP-009